requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
anthropic==0.34.2
tavily-python==0.3.9
pandas==2.2.3
aiofiles==23.2.1
httpx==0.27.2
psycopg2-binary==2.9.9
